Rename android namespace to conform to standards
