+++ android/app/src/main/java/com/mozilla/journey/MainActivity.java
@@ -1,4 +1,4 @@
-package com.magnetjourney;
+package com.mozilla.journey;
 
 import com.facebook.react.ReactActivity;
