+++ android/app/src/main/java/com/mozilla/journey/MainApplication.java
@@ -1,4 +1,4 @@
-package com.magnetjourney;
+package com.mozilla.journey;
 
 import android.app.Application;
 import android.util.Log;
